New constructor in WizardWindow to set the views
WizardWindow class should not be an abstract class. 'addWizardComponents'
could be substituted by a constructor or a set method with the list of
WizardComponents. Maybe another method to put the size of the window and
the title will be need, but this is by far more legible.

This commits introduces the posibility to set the lists in the constructor
but does not modify the previous behaviour